Add go back button.

+++ app/manage/views.py
@@ -34,7 +34,7 @@ def edit_movie(id):
             db.session.add(movie)
             db.session.commit()
             flash('电影资料已更新')
-        else:
+        elif form.delete.data:
             db.session.delete(movie)
             db.session.commit()
             flash('电影资料已删除')
@@ -54,17 +54,20 @@ def create_movie():
     form = CreateMovieForm()
     movie = Movie()
     if form.validate_on_submit():
-        movie.name = form.name.data
-        movie.date = form.date.data
-        movie.price = form.price.data
-        if form.picture.data:
-            movie.picture = request.files['picture'].read()
-        movie.director = form.director.data
-        movie.description = form.description.data
-        db.session.add(movie)
-        db.session.commit()
-        flash('电影资料已创建')
-        return redirect(url_for('manage.edit_movie', id=movie.id))
+        if form.submit.data:
+            movie.name = form.name.data
+            movie.date = form.date.data
+            movie.price = form.price.data
+            if form.picture.data:
+                movie.picture = request.files['picture'].read()
+            movie.director = form.director.data
+            movie.description = form.description.data
+            db.session.add(movie)
+            db.session.commit()
+            flash('电影资料已创建')
+            return redirect(url_for('manage.edit_movie', id=movie.id))
+        else:
+            return redirect(url_for('manage.manage_movies'))
     form.name.data = movie.name
     form.date.data = movie.date
     form.price.data = movie.price
